make code folding for GText syntax work for Sublime 4
"By default all plugins are run using Python 3.3.6, except inside the User package which always uses the latest python."
This meant that sublime_plugin.all_command_classes didn't include FoldByLevelGtextCommand.

"Any package that wishes to use Python 3.8 must create a file named .python-version in the root of the packages. This file should contain either the text 3.3 or 3.8 to select the version of Python to use."